Parallel map, single thread reduce.
This commit processes the data in N chunks across N thread where each thread returns a hashmap back to the main thread which are then all merged together in the main thread, sorted, and printed.
This update also includes the use of mmap instead of reading the file into a string, which saves about 10 seconds on an 11th gen core i7 with 2x32GB DDR4.