Replace hash key with byte slice.
In this small tweak, I shaved about 2secs off on by using the byte slice of the weather station name as the hash key instead of first parsing it to a string.
Then, at the end of the reduce phase, I parse the hashmap keys into strings for printing the results.