Use i32 instead of f64.
Down to about 8.5 seconds on the laptop. Instead of parsing each utf8 byte slice as a str and then as an f64,
parse each as a str and then parse each str as an i32 by simply ignoring the . byte. At the last step before
printing the results, convert each i32 result as an f32 by dividing by 10. This works since I didnt find
any inputs with a temp reading with greater than tens digit precision.